requirements.txt
# Core dependencies (already have some)
fastapi
uvicorn
streamlit
requests
pytest
playwright
detoxify
torch

# ollama
ollama
requests

# groqcloud
python-dotenv
pytest-html